move all distance sensor drivers to the same folder
