+++ cmake/configs/nuttx_aerofc-v1_default.cmake
@@ -18,7 +18,7 @@ set(config_module_list
 	drivers/hmc5883
 	drivers/gps
 	drivers/ist8310
-	drivers/ll40ls
+	drivers/distance_sensor
 	drivers/aerofc_adc
 	modules/sensors
