+++ cmake/configs/nuttx_auav-x21_default.cmake
@@ -20,12 +20,7 @@ set(config_module_list
 	drivers/mpu9250
 	drivers/hmc5883
 	drivers/ms5611
-	drivers/mb12xx
-	drivers/srf02
-	drivers/sf0x
-	drivers/sf1xx
-	drivers/ll40ls
-	drivers/teraranger
+	drivers/distance_sensor
 	drivers/gps
 	drivers/pwm_out_sim
 	drivers/hott
@@ -44,7 +39,6 @@ set(config_module_list
 	drivers/camera_trigger
 	drivers/bst
 	drivers/lis3mdl
-	drivers/tfmini
 
 	#
 	# System commands
@@ -71,7 +65,7 @@ set(config_module_list
 	#
 	# Testing
 	#
-	drivers/sf0x/sf0x_tests
+	drivers/distance_sensor/sf0x/sf0x_tests
 	drivers/test_ppm
 	#lib/rc/rc_tests
 	modules/commander/commander_tests
