+++ cmake/configs/nuttx_px4fmu-v5_default.cmake
@@ -28,8 +28,7 @@ set(config_module_list
 	drivers/ist8310
 	drivers/led
 	drivers/lis3mdl
-	drivers/ll40ls
-	drivers/mb12xx
+	drivers/distance_sensor
 	drivers/mkblctrl
 	drivers/mpu6000
 	drivers/mpu9250
@@ -42,17 +41,12 @@ set(config_module_list
 	drivers/px4fmu
 	drivers/rgbled
 	drivers/rgbled_pwm
-	drivers/sf0x
-	drivers/sf1xx
-	drivers/srf02
 	drivers/stm32
 	drivers/stm32/adc
 	drivers/stm32/tone_alarm
 	drivers/tap_esc
-	drivers/teraranger
 	drivers/vmount
 	modules/sensors
-	drivers/tfmini
 
 	#
 	# System commands
@@ -79,7 +73,7 @@ set(config_module_list
 	#
 	# Testing
 	#
-	drivers/sf0x/sf0x_tests
+	drivers/distance_sensor/sf0x/sf0x_tests
 	drivers/test_ppm
 	#lib/rc/rc_tests
 	modules/commander/commander_tests
