+++ cmake/configs/nuttx_px4nucleoF767ZI-v1_default.cmake
@@ -24,8 +24,7 @@ set(config_module_list
 	drivers/hott/hott_telemetry
 	drivers/led
 	drivers/lis3mdl
-	drivers/ll40ls
-	drivers/mb12xx
+	drivers/distance_sensor
 	drivers/mkblctrl
 	drivers/mpu6000
 	drivers/mpu9250
@@ -37,15 +36,11 @@ set(config_module_list
 	drivers/px4flow
 	drivers/px4fmu
 	drivers/rgbled
-	drivers/sf0x
-	drivers/srf02
 	drivers/stm32
 	drivers/stm32/adc
 	drivers/stm32/tone_alarm
 	drivers/tap_esc
-	drivers/teraranger
 	modules/sensors
-	drivers/tfmini
 
 	#
 	# System commands
