+++ src/drivers/distance_sensor/mb12xx/CMakeLists.txt
@@ -1,6 +1,6 @@
 ############################################################################
 #
-#   Copyright (c) 2017 PX4 Development Team. All rights reserved.
+#   Copyright (c) 2015 PX4 Development Team. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -31,12 +31,12 @@
 #
 ############################################################################
 px4_add_module(
-	MODULE drivers__teraranger
-	MAIN teraranger
-	STACK_MAIN 1200
+	MODULE drivers__mb12xx
+	MAIN mb12xx
 	COMPILE_FLAGS
+		-Wno-sign-compare
 	SRCS
-		teraranger.cpp
+		mb12xx.cpp
 	DEPENDS
 		platforms__common
 	)
