+++ src/drivers/distance_sensor/sf0x/sf0x_tests/CMakeLists.txt
@@ -1,6 +1,6 @@
 ############################################################################
 #
-#   Copyright (c) 2015 PX4 Development Team. All rights reserved.
+#   Copyright (c) 2016 PX4 Development Team. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -31,13 +31,13 @@
 #
 ############################################################################
 px4_add_module(
-
-	MODULE drivers__sf1xx
-	MAIN sf1xx
+	MODULE drivers__sf0x__sf0x_tests
+	MAIN sf0x_tests
 	COMPILE_FLAGS
+		-Wno-sign-compare # TODO: fix all sign-compare
 	SRCS
-		sf1xx.cpp
+		SF0XTest.cpp
+		../sf0x_parser.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix :
