+++ src/drivers/distance_sensor/sf1xx/sf1xx.cpp
@@ -1,6 +1,6 @@
 /****************************************************************************
  *
- *   Copyright (c) 2014-2015 PX4 Development Team. All rights reserved.
+ *   Copyright (c) 2013-2016 PX4 Development Team. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -32,15 +32,21 @@
  ****************************************************************************/
 
 /**
- * @file sf0x.cpp
- * @author Lorenz Meier <lm@inf.ethz.ch>
- * @author Greg Hulands
+ * @file sf1xx.cpp
  *
- * Driver for the Lightware SF0x laser rangefinder series
+ * @author ecmnet <ecm@gmx.de>
+ * @author Vasily Evseenko <svpcom@gmail.com>
+ *
+ * Driver for the Lightware SF1xx lidar range finder series.
+ * Default I2C address 0x66 is used.
  */
 
 #include <px4_config.h>
+#include <px4_defines.h>
 #include <px4_getopt.h>
+#include <px4_workqueue.h>
+
+#include <drivers/device/i2c.h>
 
 #include <sys/types.h>
 #include <stdint.h>
@@ -54,49 +60,41 @@
 #include <stdio.h>
 #include <math.h>
 #include <unistd.h>
-#include <termios.h>
-
-#include <nuttx/arch.h>
-#include <nuttx/wqueue.h>
-#include <nuttx/clock.h>
+#include <vector>
 
 #include <systemlib/perf_counter.h>
 #include <systemlib/err.h>
 
 #include <drivers/drv_hrt.h>
 #include <drivers/drv_range_finder.h>
-#include <drivers/device/device.h>
 #include <drivers/device/ringbuffer.h>
 
 #include <uORB/uORB.h>
-#include <uORB/topics/subsystem_info.h>
 #include <uORB/topics/distance_sensor.h>
 
 #include <board_config.h>
 
-#include "sf0x_parser.h"
-
 /* Configuration Constants */
+#define SF1XX_BUS 		PX4_I2C_BUS_EXPANSION
+#define SF1XX_BASEADDR 	0x66
+#define SF1XX_DEVICE_PATH	"/dev/sf1xx"
+
 
 #ifndef CONFIG_SCHED_WORKQUEUE
 # error This requires CONFIG_SCHED_WORKQUEUE.
 #endif
 
-#define SF0X_TAKE_RANGE_REG		'd'
-
-// designated SERIAL4/5 on Pixhawk
-#define SF0X_DEFAULT_PORT		"/dev/ttyS6"
-
-class SF0X : public device::CDev
+class SF1XX : public device::I2C
 {
 public:
-	SF0X(const char *port = SF0X_DEFAULT_PORT, uint8_t rotation = distance_sensor_s::ROTATION_DOWNWARD_FACING);
-	virtual ~SF0X();
+	SF1XX(uint8_t rotation = distance_sensor_s::ROTATION_DOWNWARD_FACING, int bus = SF1XX_BUS,
+	      int address = SF1XX_BASEADDR);
+	virtual ~SF1XX();
 
-	virtual int 			init();
+	virtual int 		init();
 
-	virtual ssize_t			read(struct file *filp, char *buffer, size_t buflen);
-	virtual int			ioctl(struct file *filp, int cmd, unsigned long arg);
+	virtual ssize_t		read(device::file_t *filp, char *buffer, size_t buflen);
+	virtual int			ioctl(device::file_t *filp, int cmd, unsigned long arg);
 
 	/**
 	* Diagnostics - print some basic information about the driver.
@@ -107,31 +105,32 @@ protected:
 	virtual int			probe();
 
 private:
-	char 				_port[20];
 	uint8_t _rotation;
 	float				_min_distance;
 	float				_max_distance;
-	int                 _conversion_interval;
+	int                             _conversion_interval;
 	work_s				_work;
-	ringbuffer::RingBuffer		*_reports;
+	ringbuffer::RingBuffer  *_reports;
 	bool				_sensor_ok;
 	int				_measure_ticks;
-	bool				_collect_phase;
-	int				_fd;
-	char				_linebuf[10];
-	unsigned			_linebuf_index;
-	enum SF0X_PARSE_STATE		_parse_state;
-	hrt_abstime			_last_read;
-
 	int				_class_instance;
 	int				_orb_class_instance;
 
-	orb_advert_t			_distance_sensor_topic;
+	orb_advert_t		_distance_sensor_topic;
 
-	unsigned			_consecutive_fail_count;
+	perf_counter_t		_sample_perf;
+	perf_counter_t		_comms_errors;
 
-	perf_counter_t			_sample_perf;
-	perf_counter_t			_comms_errors;
+
+
+	/**
+	* Test whether the device supported by the driver is present at a
+	* specific address.
+	*
+	* @param address	The I2C bus address to probe.
+	* @return			True if the device is present.
+	*/
+	int					probe_address(uint8_t address);
 
 	/**
 	* Initialise the automatic measurement state machine and start it.
@@ -148,8 +147,8 @@ private:
 
 	/**
 	* Set the min and max distance thresholds if you want the end points of the sensors
-	* range to be brought in at all, otherwise it will use the defaults SF0X_MIN_DISTANCE
-	* and SF0X_MAX_DISTANCE
+	* range to be brought in at all, otherwise it will use the defaults SF1XX_MIN_DISTANCE
+	* and SF1XX_MAX_DISTANCE
 	*/
 	void				set_minimum_distance(float min);
 	void				set_maximum_distance(float max);
@@ -161,8 +160,8 @@ private:
 	* and start a new one.
 	*/
 	void				cycle();
-	int				measure();
-	int				collect();
+	int					measure();
+	int					collect();
 	/**
 	* Static trampoline from the workq context; because we don't have a
 	* generic workq wrapper yet.
@@ -177,77 +176,32 @@ private:
 /*
  * Driver 'main' command.
  */
-extern "C" __EXPORT int sf0x_main(int argc, char *argv[]);
+extern "C" __EXPORT int sf1xx_main(int argc, char *argv[]);
 
-SF0X::SF0X(const char *port, uint8_t rotation) :
-	CDev("SF0X", RANGE_FINDER0_DEVICE_PATH),
+SF1XX::SF1XX(uint8_t rotation, int bus, int address) :
+	I2C("SF1XX", SF1XX_DEVICE_PATH, bus, address, 400000),
 	_rotation(rotation),
-	_min_distance(0.30f),
-	_max_distance(40.0f),
-	_conversion_interval(83334),
+	_min_distance(-1.0f),
+	_max_distance(-1.0f),
+	_conversion_interval(-1),
 	_reports(nullptr),
 	_sensor_ok(false),
 	_measure_ticks(0),
-	_collect_phase(false),
-	_fd(-1),
-	_linebuf_index(0),
-	_parse_state(SF0X_PARSE_STATE0_UNSYNC),
-	_last_read(0),
 	_class_instance(-1),
 	_orb_class_instance(-1),
 	_distance_sensor_topic(nullptr),
-	_consecutive_fail_count(0),
-	_sample_perf(perf_alloc(PC_ELAPSED, "sf0x_read")),
-	_comms_errors(perf_alloc(PC_COUNT, "sf0x_com_err"))
+	_sample_perf(perf_alloc(PC_ELAPSED, "sf1xx_read")),
+	_comms_errors(perf_alloc(PC_COUNT, "sf1xx_com_err"))
+
 {
-	/* store port name */
-	strncpy(_port, port, sizeof(_port));
-	/* enforce null termination */
-	_port[sizeof(_port) - 1] = '\0';
-
-	/* open fd */
-	_fd = ::open(_port, O_RDWR | O_NOCTTY | O_NONBLOCK);
-
-	if (_fd < 0) {
-		warnx("FAIL: laser fd");
-	}
-
-	struct termios uart_config;
-
-	int termios_state;
-
-	/* fill the struct for the new configuration */
-	tcgetattr(_fd, &uart_config);
-
-	/* clear ONLCR flag (which appends a CR for every LF) */
-	uart_config.c_oflag &= ~ONLCR;
-
-	/* no parity, one stop bit */
-	uart_config.c_cflag &= ~(CSTOPB | PARENB);
-
-	unsigned speed = B9600;
-
-	/* set baud rate */
-	if ((termios_state = cfsetispeed(&uart_config, speed)) < 0) {
-		warnx("ERR CFG: %d ISPD", termios_state);
-	}
-
-	if ((termios_state = cfsetospeed(&uart_config, speed)) < 0) {
-		warnx("ERR CFG: %d OSPD\n", termios_state);
-	}
-
-	if ((termios_state = tcsetattr(_fd, TCSANOW, &uart_config)) < 0) {
-		warnx("ERR baud %d ATTR", termios_state);
-	}
-
-	// disable debug() calls
+	/* enable debug() calls */
 	_debug_enabled = false;
 
-	// work_cancel in the dtor will explode if we don't do this...
+	/* work_cancel in the dtor will explode if we don't do this... */
 	memset(&_work, 0, sizeof(_work));
 }
 
-SF0X::~SF0X()
+SF1XX::~SF1XX()
 {
 	/* make sure we are truly inactive */
 	stop();
@@ -257,133 +211,139 @@ SF0X::~SF0X()
 		delete _reports;
 	}
 
+	if (_distance_sensor_topic != nullptr) {
+		orb_unadvertise(_distance_sensor_topic);
+	}
+
 	if (_class_instance != -1) {
 		unregister_class_devname(RANGE_FINDER_BASE_DEVICE_PATH, _class_instance);
 	}
 
+	/* free perf counters */
 	perf_free(_sample_perf);
 	perf_free(_comms_errors);
 }
 
 int
-SF0X::init()
+SF1XX::init()
 {
+	int ret = PX4_ERROR;
 	int hw_model;
-	param_get(param_find("SENS_EN_SF0X"), &hw_model);
+	param_get(param_find("SENS_EN_SF1XX"), &hw_model);
 
 	switch (hw_model) {
 	case 0:
 		DEVICE_LOG("disabled.");
-		return 0;
+		return ret;
 
-	case 1: /* SF02 (40m, 12 Hz)*/
-		_min_distance = 0.3f;
-		_max_distance = 40.0f;
-		_conversion_interval =	83334;
-		break;
-
-	case 2:  /* SF10/a (25m 32Hz) */
+	case 1:  /* SF10/a (25m 32Hz) */
 		_min_distance = 0.01f;
 		_max_distance = 25.0f;
 		_conversion_interval = 31250;
 		break;
 
-	case 3:  /* SF10/b (50m 32Hz) */
+	case 2:  /* SF10/b (50m 32Hz) */
 		_min_distance = 0.01f;
 		_max_distance = 50.0f;
 		_conversion_interval = 31250;
 		break;
 
-	case 4:  /* SF10/c (100m 16Hz) */
+	case 3:  /* SF10/c (100m 16Hz) */
 		_min_distance = 0.01f;
 		_max_distance = 100.0f;
 		_conversion_interval = 62500;
 		break;
 
-	case 5:
+	case 4:
 		/* SF11/c (120m 20Hz) */
 		_min_distance = 0.01f;
 		_max_distance = 120.0f;
 		_conversion_interval = 50000;
 		break;
 
+	case 5:
+		/* SF20/LW20 (100m 48-388Hz) */
+		_min_distance = 0.001f;
+		_max_distance = 100.0f;
+		_conversion_interval = 20834;
+		break;
+
 	default:
 		DEVICE_LOG("invalid HW model %d.", hw_model);
-		return -1;
+		return ret;
 	}
 
-	/* status */
-	int ret = 0;
+	/* do I2C init (and probe) first */
+	if (I2C::init() != OK) {
+		return ret;
+	}
 
-	do { /* create a scope to handle exit conditions using break */
+	/* allocate basic report buffers */
+	_reports = new ringbuffer::RingBuffer(2, sizeof(distance_sensor_s));
 
-		/* do regular cdev init */
-		ret = CDev::init();
+	set_device_address(SF1XX_BASEADDR);
 
-		if (ret != OK) { break; }
+	if (_reports == nullptr) {
+		return ret;
+	}
 
-		/* allocate basic report buffers */
-		_reports = new ringbuffer::RingBuffer(2, sizeof(distance_sensor_s));
+	_class_instance = register_class_devname(RANGE_FINDER_BASE_DEVICE_PATH);
 
-		if (_reports == nullptr) {
-			warnx("mem err");
-			ret = -1;
-			break;
-		}
+	/* get a publish handle on the range finder topic */
+	struct distance_sensor_s ds_report = {};
 
-		_class_instance = register_class_devname(RANGE_FINDER_BASE_DEVICE_PATH);
+	_distance_sensor_topic = orb_advertise_multi(ORB_ID(distance_sensor), &ds_report,
+				 &_orb_class_instance, ORB_PRIO_HIGH);
 
-		/* get a publish handle on the range finder topic */
-		struct distance_sensor_s ds_report = {};
+	if (_distance_sensor_topic == nullptr) {
+		DEVICE_LOG("failed to create distance_sensor object. Did you start uOrb?");
+	}
 
-		_distance_sensor_topic = orb_advertise_multi(ORB_ID(distance_sensor), &ds_report,
-					 &_orb_class_instance, ORB_PRIO_HIGH);
+	// Select altitude register
+	int ret2 = measure();
 
-		if (_distance_sensor_topic == nullptr) {
-			DEVICE_LOG("failed to create distance_sensor object. Did you start uOrb?");
-		}
+	if (ret2 == 0) {
+		ret = OK;
+		_sensor_ok = true;
+		DEVICE_LOG("(%dm %dHz) with address %d found", (int)_max_distance,
+			   (int)(1e6f / _conversion_interval), SF1XX_BASEADDR);
+	}
 
-	} while (0);
-
-	/* close the fd */
-	::close(_fd);
-	_fd = -1;
-
-	return OK;
+	return ret;
 }
 
 int
-SF0X::probe()
+SF1XX::probe()
 {
 	return measure();
 }
 
 void
-SF0X::set_minimum_distance(float min)
+SF1XX::set_minimum_distance(float min)
 {
 	_min_distance = min;
 }
 
 void
-SF0X::set_maximum_distance(float max)
+SF1XX::set_maximum_distance(float max)
 {
 	_max_distance = max;
 }
 
 float
-SF0X::get_minimum_distance()
+SF1XX::get_minimum_distance()
 {
 	return _min_distance;
 }
 
 float
-SF0X::get_maximum_distance()
+SF1XX::get_maximum_distance()
 {
 	return _max_distance;
 }
 
 int
-SF0X::ioctl(struct file *filp, int cmd, unsigned long arg)
+SF1XX::ioctl(device::file_t *filp, int cmd, unsigned long arg)
 {
 	switch (cmd) {
 
@@ -415,6 +375,7 @@ SF0X::ioctl(struct file *filp, int cmd, unsigned long arg)
 					/* if we need to start the poll state machine, do it */
 					if (want_start) {
 						start();
+
 					}
 
 					return OK;
@@ -422,12 +383,11 @@ SF0X::ioctl(struct file *filp, int cmd, unsigned long arg)
 
 			/* adjust to a legal polling interval in Hz */
 			default: {
-
 					/* do we need to start internal polling? */
 					bool want_start = (_measure_ticks == 0);
 
 					/* convert hz to tick interval via microseconds */
-					unsigned ticks = USEC2TICK(1000000 / arg);
+					int ticks = USEC2TICK(1000000 / arg);
 
 					/* check against maximum rate */
 					if (ticks < USEC2TICK(_conversion_interval)) {
@@ -460,14 +420,14 @@ SF0X::ioctl(struct file *filp, int cmd, unsigned long arg)
 				return -EINVAL;
 			}
 
-			irqstate_t flags = px4_enter_critical_section();
+			ATOMIC_ENTER;
 
 			if (!_reports->resize(arg)) {
-				px4_leave_critical_section(flags);
+				ATOMIC_LEAVE;
 				return -ENOMEM;
 			}
 
-			px4_leave_critical_section(flags);
+			ATOMIC_LEAVE;
 
 			return OK;
 		}
@@ -478,12 +438,12 @@ SF0X::ioctl(struct file *filp, int cmd, unsigned long arg)
 
 	default:
 		/* give it to the superclass */
-		return CDev::ioctl(filp, cmd, arg);
+		return I2C::ioctl(filp, cmd, arg);
 	}
 }
 
 ssize_t
-SF0X::read(struct file *filp, char *buffer, size_t buflen)
+SF1XX::read(device::file_t *filp, char *buffer, size_t buflen)
 {
 	unsigned count = buflen / sizeof(struct distance_sensor_s);
 	struct distance_sensor_s *rbuf = reinterpret_cast<struct distance_sensor_s *>(buffer);
@@ -543,19 +503,20 @@ SF0X::read(struct file *filp, char *buffer, size_t buflen)
 }
 
 int
-SF0X::measure()
+SF1XX::measure()
 {
 	int ret;
 
 	/*
-	 * Send the command to begin a measurement.
+	 * Send the command '0' -- read altitude
 	 */
-	char cmd = SF0X_TAKE_RANGE_REG;
-	ret = ::write(_fd, &cmd, 1);
 
-	if (ret != sizeof(cmd)) {
+	uint8_t cmd = 0;
+	ret = transfer(&cmd, 1, nullptr, 0);
+
+	if (OK != ret) {
 		perf_count(_comms_errors);
-		DEVICE_LOG("write fail %d", ret);
+		DEVICE_DEBUG("i2c::transfer returned %d", ret);
 		return ret;
 	}
 
@@ -565,58 +526,27 @@ SF0X::measure()
 }
 
 int
-SF0X::collect()
+SF1XX::collect()
 {
-	int	ret;
+	int	ret = -EIO;
 
+	/* read from the sensor */
+	uint8_t val[2] = {0, 0};
 	perf_begin(_sample_perf);
 
-	/* clear buffer if last read was too long ago */
-	uint64_t read_elapsed = hrt_elapsed_time(&_last_read);
-
-	/* the buffer for read chars is buflen minus null termination */
-	char readbuf[sizeof(_linebuf)];
-	unsigned readlen = sizeof(readbuf) - 1;
-
-	/* read from the sensor (uart buffer) */
-	ret = ::read(_fd, &readbuf[0], readlen);
+	ret = transfer(nullptr, 0, &val[0], 2);
 
 	if (ret < 0) {
-		DEVICE_DEBUG("read err: %d", ret);
+		DEVICE_DEBUG("error reading from sensor: %d", ret);
 		perf_count(_comms_errors);
 		perf_end(_sample_perf);
-
-		/* only throw an error if we time out */
-		if (read_elapsed > (_conversion_interval * 2)) {
-			return ret;
-
-		} else {
-			return -EAGAIN;
-		}
-
-	} else if (ret == 0) {
-		return -EAGAIN;
+		return ret;
 	}
 
-	_last_read = hrt_absolute_time();
-
-	float distance_m = -1.0f;
-	bool valid = false;
-
-	for (int i = 0; i < ret; i++) {
-		if (OK == sf0x_parser(readbuf[i], _linebuf, &_linebuf_index, &_parse_state, &distance_m)) {
-			valid = true;
-		}
-	}
-
-	if (!valid) {
-		return -EAGAIN;
-	}
-
-	DEVICE_DEBUG("val (float): %8.4f, raw: %s, valid: %s", (double)distance_m, _linebuf, ((valid) ? "OK" : "NO"));
+	uint16_t distance_cm = val[0] << 8 | val[1];
+	float distance_m = float(distance_cm) * 1e-2f;
 
 	struct distance_sensor_s report;
-
 	report.timestamp = hrt_absolute_time();
 	report.type = distance_sensor_s::MAV_DISTANCE_SENSOR_LASER;
 	report.orientation = _rotation;
@@ -627,8 +557,10 @@ SF0X::collect()
 	/* TODO: set proper ID */
 	report.id = 0;
 
-	/* publish it */
-	orb_publish(ORB_ID(distance_sensor), _distance_sensor_topic, &report);
+	/* publish it, if we are the primary */
+	if (_distance_sensor_topic != nullptr) {
+		orb_publish(ORB_ID(distance_sensor), _distance_sensor_topic, &report);
+	}
 
 	_reports->force(&report);
 
@@ -642,142 +574,70 @@ SF0X::collect()
 }
 
 void
-SF0X::start()
+SF1XX::start()
 {
 	/* reset the report ring and state machine */
-	_collect_phase = false;
 	_reports->flush();
 
+	/* set register to '0' */
+	measure();
+
 	/* schedule a cycle to start things */
-	work_queue(HPWORK, &_work, (worker_t)&SF0X::cycle_trampoline, this, 1);
-
-	// /* notify about state change */
-	// struct subsystem_info_s info = {
-	// 	true,
-	// 	true,
-	// 	true,
-	// 	SUBSYSTEM_TYPE_RANGEFINDER
-	// };
-	// static orb_advert_t pub = -1;
-
-	// if (pub > 0) {
-	// 	orb_publish(ORB_ID(subsystem_info), pub, &info);
-
-	// } else {
-	// 	pub = orb_advertise(ORB_ID(subsystem_info), &info);
-	// }
+	work_queue(HPWORK, &_work, (worker_t)&SF1XX::cycle_trampoline, this, USEC2TICK(_conversion_interval));
 }
 
 void
-SF0X::stop()
+SF1XX::stop()
 {
 	work_cancel(HPWORK, &_work);
 }
 
 void
-SF0X::cycle_trampoline(void *arg)
+SF1XX::cycle_trampoline(void *arg)
 {
-	SF0X *dev = static_cast<SF0X *>(arg);
+	SF1XX *dev = (SF1XX *)arg;
 
 	dev->cycle();
 }
 
 void
-SF0X::cycle()
+SF1XX::cycle()
 {
-	/* fds initialized? */
-	if (_fd < 0) {
-		/* open fd */
-		_fd = ::open(_port, O_RDWR | O_NOCTTY | O_NONBLOCK);
+	/* Collect results */
+	if (OK != collect()) {
+		DEVICE_DEBUG("collection error");
+		/* if error restart the measurement state machine */
+		start();
+		return;
 	}
-
-	/* collection phase? */
-	if (_collect_phase) {
-
-		/* perform collection */
-		int collect_ret = collect();
-
-		if (collect_ret == -EAGAIN) {
-			/* reschedule to grab the missing bits, time to transmit 8 bytes @ 9600 bps */
-			work_queue(HPWORK,
-				   &_work,
-				   (worker_t)&SF0X::cycle_trampoline,
-				   this,
-				   USEC2TICK(1042 * 8));
-			return;
-		}
-
-		if (OK != collect_ret) {
-
-			/* we know the sensor needs about four seconds to initialize */
-			if (hrt_absolute_time() > 5 * 1000 * 1000LL && _consecutive_fail_count < 5) {
-				DEVICE_LOG("collection error #%u", _consecutive_fail_count);
-			}
-
-			_consecutive_fail_count++;
-
-			/* restart the measurement state machine */
-			start();
-			return;
-
-		} else {
-			/* apparently success */
-			_consecutive_fail_count = 0;
-		}
-
-		/* next phase is measurement */
-		_collect_phase = false;
-
-		/*
-		 * Is there a collect->measure gap?
-		 */
-		if (_measure_ticks > USEC2TICK(_conversion_interval)) {
-
-			/* schedule a fresh cycle call when we are ready to measure again */
-			work_queue(HPWORK,
-				   &_work,
-				   (worker_t)&SF0X::cycle_trampoline,
-				   this,
-				   _measure_ticks - USEC2TICK(_conversion_interval));
-
-			return;
-		}
-	}
-
-	/* measurement phase */
-	if (OK != measure()) {
-		DEVICE_LOG("measure error");
-	}
-
-	/* next phase is collection */
-	_collect_phase = true;
 
 	/* schedule a fresh cycle call when the measurement is done */
 	work_queue(HPWORK,
 		   &_work,
-		   (worker_t)&SF0X::cycle_trampoline,
+		   (worker_t)&SF1XX::cycle_trampoline,
 		   this,
 		   USEC2TICK(_conversion_interval));
+
 }
 
 void
-SF0X::print_info()
+SF1XX::print_info()
 {
 	perf_print_counter(_sample_perf);
 	perf_print_counter(_comms_errors);
-	printf("poll interval:  %d ticks\n", _measure_ticks);
+	printf("poll interval:  %u ticks\n", _measure_ticks);
 	_reports->print_info("report queue");
 }
 
 /**
  * Local functions in support of the shell command.
  */
-namespace sf0x
+namespace sf1xx
 {
 
-SF0X	*g_dev;
+SF1XX	*g_dev;
 
-void	start(const char *port, uint8_t rotation);
+void	start(uint8_t rotation);
 void	stop();
 void	test();
 void	reset();
@@ -787,16 +647,16 @@ void	info();
  * Start the driver.
  */
 void
-start(const char *port, uint8_t rotation)
+start(uint8_t rotation)
 {
-	int fd;
+	int fd = -1;
 
 	if (g_dev != nullptr) {
 		errx(1, "already started");
 	}
 
 	/* create the driver */
-	g_dev = new SF0X(port, rotation);
+	g_dev = new SF1XX(rotation, SF1XX_BUS);
 
 	if (g_dev == nullptr) {
 		goto fail;
@@ -807,17 +667,18 @@ start(const char *port, uint8_t rotation)
 	}
 
 	/* set the poll rate to default, starts automatic data collection */
-	fd = open(RANGE_FINDER0_DEVICE_PATH, 0);
+	fd = open(SF1XX_DEVICE_PATH, O_RDONLY);
 
 	if (fd < 0) {
-		warnx("device open fail");
 		goto fail;
 	}
 
 	if (ioctl(fd, SENSORIOCSPOLLRATE, SENSOR_POLLRATE_DEFAULT) < 0) {
+		::close(fd);
 		goto fail;
 	}
 
+	::close(fd);
 	exit(0);
 
 fail:
@@ -856,11 +717,12 @@ test()
 {
 	struct distance_sensor_s report;
 	ssize_t sz;
+	int ret;
 
-	int fd = open(RANGE_FINDER0_DEVICE_PATH, O_RDONLY);
+	int fd = open(SF1XX_DEVICE_PATH, O_RDONLY);
 
 	if (fd < 0) {
-		err(1, "%s open failed (try 'sf0x start' if the driver is not running", RANGE_FINDER0_DEVICE_PATH);
+		err(1, "%s open failed (try 'sf1xx start' if the driver is not running", SF1XX_DEVICE_PATH);
 	}
 
 	/* do a simple demand read */
@@ -871,10 +733,10 @@ test()
 	}
 
 	warnx("single read");
-	warnx("measurement:  %0.2f m", (double)report.current_distance);
-	warnx("time: %llu", report.timestamp);
+	warnx("measurement: %0.2f m", (double)report.current_distance);
+	warnx("time:        %llu", report.timestamp);
 
-	/* start the sensor polling at 2 Hz rate */
+	/* start the sensor polling at 2Hz */
 	if (OK != ioctl(fd, SENSORIOCSPOLLRATE, 2)) {
 		errx(1, "failed to set 2Hz poll rate");
 	}
@@ -886,33 +748,32 @@ test()
 		/* wait for data to be ready */
 		fds.fd = fd;
 		fds.events = POLLIN;
-		int ret = poll(&fds, 1, 2000);
+		ret = poll(&fds, 1, 2000);
 
 		if (ret != 1) {
-			warnx("timed out");
-			break;
+			errx(1, "timed out waiting for sensor data");
 		}
 
 		/* now go get it */
 		sz = read(fd, &report, sizeof(report));
 
 		if (sz != sizeof(report)) {
-			warnx("read failed: got %d vs exp. %d", sz, sizeof(report));
-			break;
+			err(1, "periodic read failed");
 		}
 
-		warnx("read #%u", i);
+		warnx("periodic read %u", i);
 		warnx("valid %u", (float)report.current_distance > report.min_distance
 		      && (float)report.current_distance < report.max_distance ? 1 : 0);
-		warnx("measurement:  %0.3f m", (double)report.current_distance);
-		warnx("time: %llu", report.timestamp);
+		warnx("measurement: %0.3f", (double)report.current_distance);
+		warnx("time:        %llu", report.timestamp);
 	}
 
-	/* reset the sensor polling to the default rate */
+	/* reset the sensor polling to default rate */
 	if (OK != ioctl(fd, SENSORIOCSPOLLRATE, SENSOR_POLLRATE_DEFAULT)) {
-		errx(1, "ERR: DEF RATE");
+		errx(1, "failed to set default poll rate");
 	}
 
+	::close(fd);
 	errx(0, "PASS");
 }
 
@@ -922,7 +783,7 @@ test()
 void
 reset()
 {
-	int fd = open(RANGE_FINDER0_DEVICE_PATH, O_RDONLY);
+	int fd = open(SF1XX_DEVICE_PATH, O_RDONLY);
 
 	if (fd < 0) {
 		err(1, "failed ");
@@ -936,6 +797,7 @@ reset()
 		err(1, "driver poll restart failed");
 	}
 
+	::close(fd);
 	exit(0);
 }
 
@@ -955,16 +817,16 @@ info()
 	exit(0);
 }
 
-} // namespace
+} /* namespace */
 
 int
-sf0x_main(int argc, char *argv[])
+sf1xx_main(int argc, char *argv[])
 {
 	// check for optional arguments
 	int ch;
-	uint8_t rotation = distance_sensor_s::ROTATION_DOWNWARD_FACING;
 	int myoptind = 1;
-	const char *myoptarg = NULL;
+	const char *myoptarg = nullptr;
+	uint8_t rotation = distance_sensor_s::ROTATION_DOWNWARD_FACING;
 
 
 	while ((ch = px4_getopt(argc, argv, "R:", &myoptind, &myoptarg)) != EOF) {
@@ -983,41 +845,37 @@ sf0x_main(int argc, char *argv[])
 	 * Start/load the driver.
 	 */
 	if (!strcmp(argv[myoptind], "start")) {
-		if (argc > myoptind + 1) {
-			sf0x::start(argv[myoptind + 1], rotation);
-
-		} else {
-			sf0x::start(SF0X_DEFAULT_PORT, rotation);
-		}
+		sf1xx::start(rotation);
 	}
 
 	/*
 	 * Stop the driver
 	 */
 	if (!strcmp(argv[myoptind], "stop")) {
-		sf0x::stop();
+		sf1xx::stop();
 	}
 
 	/*
 	 * Test the driver/device.
 	 */
 	if (!strcmp(argv[myoptind], "test")) {
-		sf0x::test();
+		sf1xx::test();
 	}
 
 	/*
 	 * Reset the driver.
 	 */
 	if (!strcmp(argv[myoptind], "reset")) {
-		sf0x::reset();
+		sf1xx::reset();
 	}
 
 	/*
 	 * Print driver information.
 	 */
-	if (!strcmp(argv[myoptind], "info") || !strcmp(argv[1], "status")) {
-		sf0x::info();
+	if (!strcmp(argv[myoptind], "info") || !strcmp(argv[myoptind], "status")) {
+		sf1xx::info();
 	}
 
-	errx(1, "unrecognized command, try 'start', 'test', 'reset' or 'info'");
+	PX4_ERR("unrecognized command, try 'start', 'test', 'reset' or 'info'");
+	return PX4_ERROR;
 }
